requirements.txt
streamlit>=1.31.0
pandas>=2.2.0
plotly>=5.18.0
numpy>=1.24.0

anthropic>=0.18.0
google-generativeai>=0.3.0
google-genai>=0.1.0

datasets>=2.16.0
pydantic>=2.5.0
python-dotenv>=1.0.0

kaleido>=0.2.1
statsmodels>=0.14.0
